requirements.txt
google-cloud-storage
google-cloud-speech
crewai
crewai-tools
pydantic
streamlit
